refactor: update import paths for Movie type in adapters and API service for consistency
