+++ src/core/movie/services/adapter/get-movie-adapter.ts
@@ -1,4 +1,4 @@
-import { Movie } from '../types/movie';
+import { Movie } from '../../types/movie';
 
 
 export const getMovieDetailsAdapter = (movie: any): Movie => {
